# Speed up joins on indexed tables with string keys