@gianm 2020-01-29
How do these numbers compare to the standard lookup case?