@ccaominh 2020-01-29
> How do these numbers compare to the standard lookup case?

Lookups are about 5x faster:

```
Benchmark                               (columnCacheSizeBytes)  Score   Error  Units
lookupVirtualColumnStringKey                                 0  3.374 ± 0.065  ms/op
lookupVirtualColumnStringKey                             16384  3.364 ± 0.054  ms/op
lookupVirtualColumnStringKeyWithFilter                       0  4.553 ± 0.114  ms/op
lookupVirtualColumnStringKeyWithFilter                   16384  5.529 ± 0.305  ms/op
```