@ccaominh 2020-02-01
Benchmark numbers with the improved cache implementation:

#### Original
```
Benchmark                               (columnCacheSizeBytes)   Score   Error  Units
joinIndexedTableStringKey                                    0  42.339 ± 0.323  ms/op
joinIndexedTableStringKey                                16384  22.634 ± 0.247  ms/op
joinIndexedTableStringKeyWithFilter                          0  38.170 ± 0.268  ms/op
joinIndexedTableStringKeyWithFilter                      16384  24.974 ± 0.462  ms/op
lookupVirtualColumnStringKey                                 0   4.040 ± 0.042  ms/op
lookupVirtualColumnStringKey                             16384   3.194 ± 0.011  ms/op
lookupVirtualColumnStringKeyWithFilter                       0   5.288 ± 0.027  ms/op
lookupVirtualColumnStringKeyWithFilter                   16384   5.175 ± 0.042  ms/op
```

#### Int2IntListLruCache
```
Benchmark                               (columnCacheSizeBytes)   Score   Error  Units
joinIndexedTableStringKey                                    0  14.557 ± 0.151  ms/op
joinIndexedTableStringKey                                16384  15.167 ± 0.075  ms/op
joinIndexedTableStringKeyWithFilter                          0  16.093 ± 0.517  ms/op
joinIndexedTableStringKeyWithFilter                      16384  16.434 ± 0.185  ms/op
```

#### Int2IntListLookupTable
```
Benchmark                               (columnCacheSizeBytes)   Score   Error  Units
joinIndexedTableStringKey                                    0  12.759 ± 0.125  ms/op
joinIndexedTableStringKey                                16384  12.439 ± 0.481  ms/op
joinIndexedTableStringKeyWithFilter                          0  14.139 ± 0.127  ms/op
joinIndexedTableStringKeyWithFilter                      16384  14.163 ± 0.176  ms/op
```